将ConnectID由64位，统一改成了32位
还有一些代码调整.

+++ project/Server/ServerEngine/Connection.h
@@ -110,7 +110,7 @@ public:
 
 	VOID		    DeleteConnection(CConnection *pConnection);
 
-    CConnection*    GetConnectionByConnID(UINT64 dwConnID);
+    CConnection*    GetConnectionByConnID(UINT32 dwConnID);
 
 	///////////////////////////////////////////
 	BOOL		    CloseAllConnection();
